Make tests green and get rid of warnings
This also unifies clj/cljs code in some places, like using an `array?` check
instead of relying on object type.

Also clean up use of extend/extend-protocol. The original version used `extend` for
nil and Object out of a mistaken belief that `extend-protocol` wouldn't work for
these cases, but it works fine. For ClojureScript the default value is not
`Object` or `js/Object` but `default`.

- Fix diffing of records
- Make sure type-name always returns a symbol
- Fix printing of time values
- Make the registry feature availabe cross platform by using an atom instead of vars