Prepare for the release of the CLJC deep-diff2
Move namespaces to use a "2" suffix, we have breaking changes and we don't want
to break existing consumers. We will bundle both namespaces with and without the
suffix in the released artifact.

Move the printer and diff namespaces to printer-impl and diff-impl, for two
reasons. This way we can keep the same top level API at
least (lambdaisland.deep-diff2/printer, lambdaisland.deep-diff2/diff), since in
ClojureScript names of functions and namespaces are not allowed to overlap.

The second reason is that the -impl suffix signals that these are internal and
basically have to be treated as private. This allows us to make the functions in
there to be public, which makes it possible to unit test them. Before we used
alter-meta! to make them public at the start of the test, but this only works in
Clojure.

Inline Puget, we've made extensive changes to make this work, which we don't
expect to go upstream, so we should not release these under puget.* namespaces.

Make Puget dispatch based on value rather than type for both Clojure and
ClojureScript. Clean up and further unify the print handlers for
platform-specific types.

Contains a fix to the round-trip generative test, since ##NaN (on js/cljs) can
not be compared for equality it leads to false negatives.